extract position and range to separate files
